Calculate title from AIP folder instead of argument
Using an argument is extra work for staff and requires making sure the exact formatting of the department name is provided. Hargrett and Russell AIPs have a consistent prefix, so use that to test.

Added bonus that could process a mixed batch of AIPs if desired, as long as update how the manifests are made to split the two.