Make log a function
It repeats in a few places, so even though it is only a few lines, it is easier to keep up to date as a function.

It also makes it easy to start the log before looping through the script to add the header row instead of checking for if it has a header row each time or tracking if it is the first AIP. Added to items to skip since it is now in the AIPs directory prior to starting the loop.