Only check extension of first file to determine type
The code before checked every file but updated the value for aip_type each time so just the value of the last file was saved. It didn't really check for mixed AIPs (an error). To speed up the code, break out of the loop as soon as one file is checked.