Rename AIP folder for Hargrett in package(), like Russell
Originally renamed Hargrett AIPs from ID_title to ID as soon as the title was extracted, like the general AIP script. However, if there is a media and a metadata AIP for the same AIP, that would result in two files with the same folder titles.

Instead. leave the unique name of id_title until package(), which is when the script renames all aips to id_type_bag. The only other change is to test for if aip_folder is still in the directory instead of aip_id to see if it has been moved.