Fixing error handling in package()
Removed --quiet from bagit.py --validate because quiet caused all validation outputs to be empty, even if they were invalid.

Changed the name of the text file with the error output to include the bag name instead of AIP ID so if the metadata and media for the same AIP are included in the same batch, their bag validation errors are kept as separate documents.