Add Script Input to README
To make it easier for people new to the workflow to get oriented, added a section explaining how the AIPs Directory should be organized prior to running the script, with additional detail on Hargrett.

Deleted the explanation of the AIPs directory from  Script Usage since it duplicated what was in Script Input.