Pass AIP ID to stylesheet as parameter
Previously, the stylesheet calculated the AIP ID from the MediaInfo XML. This output changes sometimes, which results in bugs when the correct id is no longer extracted.

To avoid this, passing the AIP ID as a parameter to the stylesheet instead. It has already been calculated by the Python script and is already an argument for the preservation_xml() function.