Use version from general track
We started preserving the Matroska format, which has format_version in both the general and video tracks in the MediaInfo output. Having multiple versions breaks the concat used to remove duplicates in the AIP section. Upon further analysis of the formats in use, we determined that the only relevant version information to include is from the general track.